refactor: 배경 이미지 오버레이 컴포넌트 분리. titleSection 컴포넌트 분리. #58
