+++ src/features/freeTrial/analyzeProcess.js
@@ -2,44 +2,22 @@ import React from "react";
 import FirstStep from "./firstStep";
 import FeatureCardGrid from "./components/featureCardGrid";
 import UploadBox from "./components/uploadBox";
+import TitleSection from "./components/titleSection";
+import BackGroundOverlay from "./components/backGroundOverlay";
 export default function analyzeProcess() {
-    const rawSvg = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 800 800"><defs><linearGradient x1="50%" y1="0%" x2="50%" y2="100%" id="ffflurry-grad" gradientTransform="rotate(270)"><stop stop-color="hsl(0, 0%, 100%)" stop-opacity="1" offset="0%"></stop><stop stop-color="hsl(212, 42%, 76%)" stop-opacity="1" offset="45%"></stop><stop stop-color="hsl(167, 14%, 86%)" stop-opacity="1" offset="100%"></stop></linearGradient></defs><rect width="100%" height="100%" fill="hsl(0, 0%, 100%)"></rect><g fill="url(#ffflurry-grad)"><rect width="65" height="1" x="464.5" y="332.5" rx="0.5" transform="rotate(45, 497, 333)" opacity="0.54"></rect><rect width="270" height="1" x="-74" y="613.5" rx="0.5" transform="rotate(45, 61, 614)" opacity="0.73"></rect><rect width="463" height="1" x="-182.5" y="84.5" rx="0.5" transform="rotate(45, 49, 85)" opacity="0.71"></rect><rect width="671" height="1" x="-286.5" y="266.5" rx="0.5" transform="rotate(45, 49, 267)" opacity="0.86"></rect><rect width="68" height="1" x="222" y="413.5" rx="0.5" transform="rotate(45, 256, 414)" opacity="0.43"></rect><rect width="189" height="1" x="673.5" y="569.5" rx="0.5" transform="rotate(45, 768, 570)" opacity="0.21"></rect><rect width="286" height="1" x="612" y="386.5" rx="0.5" transform="rotate(45, 755, 387)" opacity="0.17"></rect><rect width="281" height="1" x="464.5" y="744.5" rx="0.5" transform="rotate(45, 605, 745)" opacity="0.49"></rect><rect width="48" height="1" x="522" y="233.5" rx="0.5" transform="rotate(45, 546, 234)" opacity="0.90"></rect><rect width="185" height="1" x="636.5" y="106.5" rx="0.5" transform="rotate(45, 729, 107)" opacity="0.30"></rect><rect width="270" height="1" x="25" y="622.5" rx="0.5" transform="rotate(45, 160, 623)" opacity="0.38"></rect><rect width="206" height="1" x="72" y="89.5" rx="0.5" transform="rotate(45, 175, 90)" opacity="0.41"></rect><rect width="72" height="1" x="3" y="485.5" rx="0.5" transform="rotate(45, 39, 486)" opacity="0.11"></rect><rect width="168" height="1" x="478" y="642.5" rx="0.5" transform="rotate(45, 562, 643)" opacity="0.69"></rect><rect width="583" height="1" x="83.5" y="722.5" rx="0.5" transform="rotate(45, 375, 723)" opacity="0.36"></rect><rect width="126" height="1" x="341" y="101.5" rx="0.5" transform="rotate(45, 404, 102)" opacity="0.11"></rect><rect width="187" height="1" x="258.5" y="541.5" rx="0.5" transform="rotate(45, 352, 542)" opacity="0.26"></rect><rect width="469" height="1" x="512.5" y="749.5" rx="0.5" transform="rotate(45, 747, 750)" opacity="0.80"></rect><rect width="325" height="1" x="92.5" y="477.5" rx="0.5" transform="rotate(45, 255, 478)" opacity="0.08"></rect><rect width="457" height="1" x="173.5" y="496.5" rx="0.5" transform="rotate(45, 402, 497)" opacity="0.53"></rect><rect width="122" height="1" x="394" y="659.5" rx="0.5" transform="rotate(45, 455, 660)" opacity="0.18"></rect><rect width="345" height="1" x="510.5" y="303.5" rx="0.5" transform="rotate(45, 683, 304)" opacity="0.28"></rect><rect width="279" height="1" x="-95.5" y="389.5" rx="0.5" transform="rotate(45, 44, 390)" opacity="0.14"></rect><rect width="174" height="1" x="390" y="443.5" rx="0.5" transform="rotate(45, 477, 444)" opacity="0.61"></rect><rect width="233" height="1" x="45.5" y="156.5" rx="0.5" transform="rotate(45, 162, 157)" opacity="0.62"></rect><rect width="324" height="1" x="44" y="27.5" rx="0.5" transform="rotate(45, 206, 28)" opacity="0.96"></rect><rect width="708" height="1" x="327" y="579.5" rx="0.5" transform="rotate(45, 681, 580)" opacity="0.88"></rect><rect width="548" height="1" x="-157" y="456.5" rx="0.5" transform="rotate(45, 117, 457)" opacity="0.41"></rect><rect width="389" height="1" x="181.5" y="383.5" rx="0.5" transform="rotate(45, 376, 384)" opacity="0.93"></rect><rect width="380" height="1" x="40" y="162.5" rx="0.5" transform="rotate(45, 230, 163)" opacity="0.40"></rect><rect width="410" height="1" x="-48" y="741.5" rx="0.5" transform="rotate(45, 157, 742)" opacity="1.00"></rect><rect width="433" height="1" x="-28.5" y="429.5" rx="0.5" transform="rotate(45, 188, 430)" opacity="0.52"></rect><rect width="164" height="1" x="393" y="198.5" rx="0.5" transform="rotate(45, 475, 199)" opacity="0.65"></rect><rect width="165" height="1" x="61.5" y="357.5" rx="0.5" transform="rotate(45, 144, 358)" opacity="0.47"></rect><rect width="355" height="1" x="140.5" y="655.5" rx="0.5" transform="rotate(45, 318, 656)" opacity="0.96"></rect><rect width="209" height="1" x="658.5" y="67.5" rx="0.5" transform="rotate(45, 763, 68)" opacity="0.49"></rect><rect width="213" height="1" x="136.5" y="644.5" rx="0.5" transform="rotate(45, 243, 645)" opacity="0.48"></rect><rect width="59" height="1" x="241.5" y="111.5" rx="0.5" transform="rotate(45, 271, 112)" opacity="0.66"></rect><rect width="332" height="1" x="141" y="580.5" rx="0.5" transform="rotate(45, 307, 581)" opacity="0.28"></rect><rect width="67" height="1" x="15.5" y="722.5" rx="0.5" transform="rotate(45, 49, 723)" opacity="0.68"></rect><rect width="388" height="1" x="216" y="595.5" rx="0.5" transform="rotate(45, 410, 596)" opacity="0.86"></rect><rect width="252" height="1" x="418" y="86.5" rx="0.5" transform="rotate(45, 544, 87)" opacity="0.59"></rect><rect width="413" height="1" x="96.5" y="202.5" rx="0.5" transform="rotate(45, 303, 203)" opacity="0.76"></rect><rect width="458" height="1" x="368" y="284.5" rx="0.5" transform="rotate(45, 597, 285)" opacity="0.46"></rect><rect width="65" height="1" x="732.5" y="221.5" rx="0.5" transform="rotate(45, 765, 222)" opacity="0.14"></rect><rect width="381" height="1" x="195.5" y="176.5" rx="0.5" transform="rotate(45, 386, 177)" opacity="0.55"></rect><rect width="325" height="1" x="11.5" y="269.5" rx="0.5" transform="rotate(45, 174, 270)" opacity="0.24"></rect><rect width="303" height="1" x="432.5" y="531.5" rx="0.5" transform="rotate(45, 584, 532)" opacity="0.30"></rect><rect width="67" height="1" x="555.5" y="404.5" rx="0.5" transform="rotate(45, 589, 405)" opacity="0.54"></rect><rect width="162" height="1" x="174" y="44.5" rx="0.5" transform="rotate(45, 255, 45)" opacity="0.84"></rect><rect width="138" height="1" x="245" y="38.5" rx="0.5" transform="rotate(45, 314, 39)" opacity="0.61"></rect><rect width="149" height="1" x="687.5" y="17.5" rx="0.5" transform="rotate(45, 762, 18)" opacity="0.45"></rect><rect width="316" height="1" x="426" y="160.5" rx="0.5" transform="rotate(45, 584, 161)" opacity="0.43"></rect><rect width="650" height="1" x="-44" y="750.5" rx="0.5" transform="rotate(45, 281, 751)" opacity="0.17"></rect><rect width="135" height="1" x="590.5" y="37.5" rx="0.5" transform="rotate(45, 658, 38)" opacity="0.72"></rect><rect width="389" height="1" x="53.5" y="555.5" rx="0.5" transform="rotate(45, 248, 556)" opacity="0.92"></rect><rect width="161" height="1" x="278.5" y="603.5" rx="0.5" transform="rotate(45, 359, 604)" opacity="0.68"></rect><rect width="704" height="1" x="34" y="278.5" rx="0.5" transform="rotate(45, 386, 279)" opacity="0.73"></rect><rect width="418" height="1" x="-109" y="194.5" rx="0.5" transform="rotate(45, 100, 195)" opacity="0.99"></rect><rect width="79" height="1" x="-16.5" y="556.5" rx="0.5" transform="rotate(45, 23, 557)" opacity="0.45"></rect><rect width="32" height="1" x="286" y="507.5" rx="0.5" transform="rotate(45, 302, 508)" opacity="0.25"></rect><rect width="94" height="1" x="523" y="25.5" rx="0.5" transform="rotate(45, 570, 26)" opacity="0.40"></rect><rect width="444" height="1" x="262" y="554.5" rx="0.5" transform="rotate(45, 484, 555)" opacity="0.71"></rect><rect width="454" height="1" x="-121" y="537.5" rx="0.5" transform="rotate(45, 106, 538)" opacity="0.76"></rect><rect width="525" height="1" x="11.5" y="338.5" rx="0.5" transform="rotate(45, 274, 339)" opacity="0.72"></rect><rect width="364" height="1" x="-55" y="29.5" rx="0.5" transform="rotate(45, 127, 30)" opacity="0.17"></rect><rect width="199" height="1" x="375.5" y="753.5" rx="0.5" transform="rotate(45, 475, 754)" opacity="1.00"></rect><rect width="232" height="1" x="218" y="455.5" rx="0.5" transform="rotate(45, 334, 456)" opacity="0.89"></rect><rect width="404" height="1" x="450" y="116.5" rx="0.5" transform="rotate(45, 652, 117)" opacity="0.40"></rect><rect width="439" height="1" x="250.5" y="62.5" rx="0.5" transform="rotate(45, 470, 63)" opacity="0.15"></rect><rect width="51" height="1" x="363.5" y="28.5" rx="0.5" transform="rotate(45, 389, 29)" opacity="0.73"></rect><rect width="226" height="1" x="220" y="115.5" rx="0.5" transform="rotate(45, 333, 116)" opacity="0.40"></rect><rect width="125" height="1" x="122.5" y="519.5" rx="0.5" transform="rotate(45, 185, 520)" opacity="0.61"></rect><rect width="365" height="1" x="518.5" y="677.5" rx="0.5" transform="rotate(45, 701, 678)" opacity="0.44"></rect><rect width="319" height="1" x="524.5" y="462.5" rx="0.5" transform="rotate(45, 684, 463)" opacity="0.73"></rect><rect width="549" height="1" x="406.5" y="196.5" rx="0.5" transform="rotate(45, 681, 197)" opacity="0.92"></rect><rect width="491" height="1" x="11.5" y="258.5" rx="0.5" transform="rotate(45, 257, 259)" opacity="0.17"></rect></g></svg>`
-    const encoded = encodeURIComponent(rawSvg);
+
 
     return (
-        <div
-            style={{
-                position: 'absolute',
-                inset: 0,
-                width: '100%',
-                height: '100%',
-                background: `url("data:image/svg+xml;utf8,${encoded}") center/cover no-repeat`
-            }}
-        >
-        <div className="mt-4 mx-10">
-            <div className="grid grid-cols-1 grid-rows-1 gap-10">
-                <div className="">
-                    <div className="items-center justify-center gap-2 mt-8">
-                        <h1 className="text-center font-extrabold tracking-tight text-transparent text-4xl sm:text-4xl xl:text-5xl text-slate-600 h-full">
-                            Manipulated Image Detection
-                        </h1>
-
-                        <h2 className="text-gray-600 font-extrabold text-4xl tracking-tight text-center ">
-                            Manipulated Image Detection
-                        </h2>
-
-                        <p className="text-xs font-light text-gray-600 tracking-tight text-center mt-2">
-                            Please upload an image that you find suspicious.
-                        </p>
-                    </div>
-
+        <BackGroundOverlay>
+            <div className="mt-4 mx-10">
+                <div className="grid grid-cols-1 grid-rows-1 gap-10">
                     <div>
+                        <TitleSection/>
                         <UploadBox/>
                     </div>
+                    <FeatureCardGrid/>
                 </div>
-                <FeatureCardGrid/>
             </div>
-        </div>
-        </div>
+        </BackGroundOverlay>
     );
 }
